DATA-MYSQL/pyproject.toml
[project]
name = "data-mysql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql-connector>=2.2.9",
    "sqlalchemy>=2.0.44",
]
